chore: the code was restructured according to the mvc template
