+++ main.py
@@ -1,15 +1,13 @@
+from utils.model import users
+from utils.controller import get_user_info
 
 
-users:list=[
-    {'name':'Maciej','location':'Łódź','posts':100},
-    {'name':'Mateusz','location':'Poznań','posts':200},
-    {'name':'Maciej1','location':'Siedlce','posts':300},
-    {'name':'Konrad','location':'Grudziądz','posts':400},
-]
 
 
-def get_user_info(users_data:list)->None:
-    for user in users_data:
-     print(f'twój znajomy {user["name"]} z miejscowości {user["location"]} opublikował {user["posts"]} postów')
+def main():
+    get_user_info(users)
 
-get_user_info(users)
+
+
+if __name__ == '__main__':
+    main()
